Move the folder of the images to the assets folder
