+++ src/assets/images/model-generation-app/how_it_works.png
