+++ src/assets/images/model-generation-app/model_generation_app_preview.gif
